Clean Up

+++ server/ACADEMY.Application/StorageService/StorageService.cs
@@ -6,26 +6,22 @@ namespace ACADEMY.Application.StorageService
 {
     public class StorageService : IStorageService
     {
+        private const string ContentFolderName = "Contents";
         private readonly string _contentFolder;
-        
-        private const string CONTENT_FOLDER_NAME = "Contents";
 
-        public StorageService(IWebHostEnvironment  webHostEnvironment)
+        public StorageService(IWebHostEnvironment webHostEnvironment)
         {
-            _contentFolder = Path.Combine(webHostEnvironment.WebRootPath, CONTENT_FOLDER_NAME);
+            _contentFolder = Path.Combine(webHostEnvironment.WebRootPath, ContentFolderName);
         }
-        
+
         public async Task SaveFileAsync(Stream mediaBinaryStream, string directory, string fileName)
         {
             var directoryPath = Path.Combine(_contentFolder, directory);
-            
-            if (!Directory.Exists(directoryPath))
-            {
-                Directory.CreateDirectory(directoryPath);
-            }
+
+            if (!Directory.Exists(directoryPath)) Directory.CreateDirectory(directoryPath);
 
             var filePath = Path.Combine(directoryPath, fileName);
-            
+
             await using var output = new FileStream(filePath, FileMode.Create);
             await mediaBinaryStream.CopyToAsync(output);
         }
@@ -33,10 +29,7 @@ namespace ACADEMY.Application.StorageService
         public async Task DeleteFileAsync(string directory, string fileName)
         {
             var filePath = Path.Combine(_contentFolder, directory, fileName);
-            if (File.Exists(filePath))
-            {
-                await Task.Run(() => File.Delete(filePath));
-            }
+            if (File.Exists(filePath)) await Task.Run(() => File.Delete(filePath));
         }
 
         public Task<string> GetFilePathAsync(string directory, string fileName)
